agent-requirements.txt
# LangGraph and AI Dependencies
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
pydantic>=2.5.0

# Async Support
aiohttp>=3.9.0

# Meta Ads SDK
facebook-business>=18.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0